requirements.txt
streamlit>=1.32.0
huggingface_hub>=0.23.0
pillow>=10.2.0
